+++ pkg/coordinator/steps/utils.go
@@ -23,6 +23,7 @@ import (
 	"github.com/go-logr/logr"
 
 	logutil "github.com/llm-d/llm-d-router/pkg/common/observability/logging"
+	reqcommon "github.com/llm-d/llm-d-router/pkg/common/request"
 
 	"github.com/llm-d/llm-d-router/pkg/coordinator/gateway"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/pipeline"
@@ -69,6 +70,40 @@ func resolveFormat(useOpenAIFormat bool, path string) gateway.RequestFormat {
 		return gateway.FormatGenerate
 	}
 	return detected
+}
+
+// capSingleTokenOutput rewrites body into a single-output-token, non-streaming
+// request for the synthetic prefill and encode legs.
+//
+// Intentionally distinct from the sidecar's reqcommon.PrimeSingleTokenRequest
+// for now.
+// TODO: unify the two into one shared single-token helper in a future refactor.
+func capSingleTokenOutput(body map[string]any, format gateway.RequestFormat) {
+	target := body
+	if format == gateway.FormatGenerate {
+		sp, ok := body[reqcommon.FieldSamplingParams].(map[string]any)
+		if !ok {
+			sp = map[string]any{}
+			body[reqcommon.FieldSamplingParams] = sp
+		}
+		target = sp
+	}
+
+	target[reqcommon.FieldMaxTokens] = 1
+	// Strip rather than clamp min_tokens: it defaults to 0 in vLLM, so removing it
+	// keeps min_tokens <= max_tokens=1 without raising the floor above the cap.
+	delete(target, reqcommon.FieldMinTokens)
+
+	if _, ok := body[reqcommon.FieldMaxCompletionTokens]; ok {
+		body[reqcommon.FieldMaxCompletionTokens] = 1
+	}
+
+	// TODO: max_output_tokens is another client-supplied output cap (Responses
+	// API) that a client can send instead of max_tokens/max_completion_tokens; it
+	// should be capped to 1 here as well so the synthetic legs stay single-token.
+
+	body[reqcommon.FieldStream] = false
+	delete(body, reqcommon.FieldStreamOptions)
 }
 
 // buildMMFeatures builds the multimodal features map (mm_hashes, mm_placeholders,
